+++ core/src/main/java/com/adchain/AdChainListener.java
@@ -5,9 +5,14 @@ import android.support.annotation.MainThread;
 /**
  * Created by Gust on 19.12.2017.
  */
-public interface AdChainListener {
-    //    void lastAdClosed();
-//    void penultimateAdClosed();
+public class AdChainListener {
+
     @MainThread
-    void adCompleted(int order, int total, boolean isLastAd);
+    void adCompleted(int order, int total, boolean isLastAd){}
+
+    @MainThread
+    void adClicked(Class<? extends AdChainAdapter> type){}
+
+    @MainThread
+    void adLoaded(Class<? extends AdChainAdapter> type){}
 }