+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+
+## [0.8.0] - 2018-10-07
+### Added
+- Ad Loaded callback
+- Ad Clicked callback
+
+
 ## [0.7.2] - 2018-07-30
 ### Added
 - Applovin and Startapp support.